+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -1,6 +1,7 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.List;
+import org.apache.coyote.http11.HttpMethod;
 
 public class HttpRequest {
