+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestProcessor.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
+
+import org.apache.coyote.http11.HttpResponse;
 
 public interface HttpRequestProcessor {
