+++ tomcat/src/main/java/org/apache/coyote/http11/request/RootRequestProcessor.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
+
+import org.apache.coyote.http11.HttpResponse;
 
 public class RootRequestProcessor implements HttpRequestProcessor{
     @Override
