+++ tomcat/src/test/java/org/apache/coyote/http11/HttpRequestTest.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
-import org.assertj.core.api.Assertions;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
